requirements.txt
openai
fastapi
uvicorn
python-multipart
numpy
websockets<14
scipy
google-generativeai
pytest
pytest-asyncio
pytest-mock
httpx
pydub
python-dotenv